Add files via upload
This function is useful if you are bolding values and want to relate the bold value to another cell on a different sheet or workbook. I needed this because there were values that can not be chosen simply based off of values (A human has to make the final decision which may not be what a formula would deem the best option).