Add files via upload
This is a sub I use to move files, usually before creating a new file and saving. In the actual module I have constants for the toRange and fromRange but VBA will not let you use actual Ranges as constants so I just use their string values instead. You could easitly change this to take in an actual range value. The pathforNewfile is where the new file will be saved after all of the old files currently in that directory are moved to the pathForOldFile. The fileNameExpression needs to start with a \ , eg. "\companyNeeds*.xlsx" - note you can use the wildcard "*" character in this expression. This will find all files that match this pattern. Before moving the files we check to see if a file already exists with that name and rename if so. the 'FileOrDirExists' function is also in this repo separately.