Add files via upload
I wrap this sub in another sub for a button to call that will popup a File Dialog box for the user to pick a directory. The cell range passed to the TargetCell parameter is used to set the value of the related cell. This value is used by my sub that moves files and also by some subs that save files after generating and/or renaming.