Add files via upload
Pretty basic but took me a while to find the syntax and functionality I wanted. This sub assumes there is a cell on the workbook containing the code that holds the email addresses (either comma or semicolon separated) for the 'to' line of the email. This sub checks the string in the toRange cell for commas and replaces them with semicolons so outlook understands. This sub also assumes at least one attachment but only up to two can be attached. More can easily be added if desired although you sort of need to know what the max ammount of attachments you want to allow with this code. You could use parsing or take in an array for attachments instead to make the code more dynamic. This new email window will be displayed for the user to check before sending the email.