Add files via upload
This is designed to be used after a path has been generated for a file (Usually by adding a date or time to a base name). This checks to see if the filename already exists and if it does, the file is renamed by adding "(#)" before the extension where # = the number of times the file was renamed. Note. This function uses the  FileOrDirExists function (Uploaded as "CheckForFileOrDirectory").